% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_Pmetrics}
\alias{install_Pmetrics}
\title{Install Pmetrics from Github.}
\usage{
install_Pmetrics(ref = "HEAD", force = FALSE, ask = FALSE)
}
\arguments{
\item{ref}{The branch to install from. Default is "HEAD", which installs the current
release. Choose another branch like "dev" for access to beta versions.}

\item{force}{Re-install even if installed version is the same as repository version.
Default is \code{FALSE}.}

\item{ask}{Ask before installing. Default is \code{FALSE}.}
}
\description{
Installs Pmetrics from the \href{https://github.com/LAPKB/Pmetrics}{LAPKB Github repository}.
}
