% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{launch}
\alias{launch}
\title{Launch tutorials in the package}
\usage{
launch(browser = TRUE)
}
\arguments{
\item{browser}{Launch in browser, rather than Rstudio window. Default is \code{TRUE}.}
}
\description{
Provides a list of current tutorials and asks user which one to launch.
}
