% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidyup}
\alias{tidyup}
\title{Clean text}
\usage{
tidyup(x)
}
\arguments{
\item{x}{the string to process}
}
\description{
Remove spaces, leading zeros, trailing zeros after decimals and numbers and trailing zeros after decimals
}
